requirements.txt
altair==5.5.0
app==0.0.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
filelock==3.17.0
fonttools==4.55.8
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.27.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pra==1.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
snscrape==0.7.0.20230622
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.0
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
vaderSentiment==3.3.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
yf==0.0.4
yfinance==0.2.52
